pyproject.toml
[project]
name = "chat-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "faiss-cpu>=1.9.0",
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "google-generativeai>=0.8.4",
    "httpx>=0.28.1",
    "python-jose>=3.3.0",
    "langchain>=1.2.12",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=2.0.8",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.11",
    "numpy>=1.26.4",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.9.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.12",
    "scalar-fastapi>=1.8.0",
    "sentence-transformers>=3.4.1",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
    "pymupdf>=1.27.2",
]
